move remotes loader into loaders package
